+++ library/math/n_choose_k/large_n_choose_small_k.hpp
@@ -1,6 +1,6 @@
 /** @file */
 #pragma once
-const int mod = 17; /**< must be prime */
+const ll mod = 17; /**< must be prime */
 /**
  * @param n arbitrarily large integer
  * @param k integer with k < mod
